type 3 works, but no rotations yet

+++ parse_cli_opts.cpp
@@ -55,7 +55,7 @@ int parse_cli_opts(int argc, char *argv[], parameters &params) {
     po::options_description opt_type3(
                 "Options for type 3 simulation, balls in a container");
     opt_type2.add_options()
-        ("N", po::value<double>(&(params.N)),
+        ("N", po::value<int>(&(params.N)),
             "number of balls")
         ("Eavg", po::value<double>(&(params.Eavg)),
             "balls' mean starting velocity [m s^-1]")
